gitlab/gitlab.go
package gitlab

import (
	"encoding/json"
	"fmt"
	"io/ioutil"
	"net/http"
	"os"
)

// Repo represents a GitLab repository.
type Repo struct {
	ID          int    `json:"id"`
	Name        string `json:"name"`
	Description string `json:"description"`
	URL         string `json:"url"`
}

func GetRepos(projectID int) ([]Repo, error) {
	client := &http.Client{}

	url := fmt.Sprintf("https://gitlab.com/api/v4/groups/%d/projects", projectID)
	method := "GET"
	var bearer = "Bearer " + os.Getenv("GITLAB_ACCESS_TOKEN")

	req, err := http.NewRequest(method, url, nil)
	if err != nil {
		return nil, err
	}
	req.Header.Add("Authorization", bearer)

	if err != nil {
		return nil, err
	}

	resp, err := client.Do(req)
	if err != nil {
		return nil, err
	}
	defer resp.Body.Close()

	body, err := ioutil.ReadAll(resp.Body)
	if err != nil {
		return nil, err
	}

	var repos []Repo
	err = json.Unmarshal(body, &repos)
	if err != nil {
		return nil, err
	}

	return repos, nil
}
